refactor: sponsors 데이터 파일 위치 변경 (components -> data)
- `sponsors.ts`를 `componets/2025-commingsoon/constants.ts` -> `data/2025`로 이동
