+++ data/2025/sponsors.ts
